chore(theme-generator): update filenames and re-export createTheme
